cập nhật: SearchService áp dụng ElasticSearch

+++ CommonService/src/main/java/com/luanvan/commonservice/model/response/AllProductResponseModel.java
@@ -99,8 +99,8 @@ public class AllProductResponseModel {
     public static class ProductVariant {
         private String id;
         private Size size;
-        private Integer stock;
-        private Integer sold;
+        private Long stock;
+        private Long sold;
         private Boolean isActive;
     }
